+++ Project/gearworks/src/include/gwassets.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "../lib/rendering/textures/Texture2D.h"