+++ Project/gearworks/src/include/gwrendering.h
@@ -4,7 +4,9 @@
 
 #include "../lib/rendering/gl_buffer_objs/vbo.hpp"
 #include "../lib/rendering/gl_buffer_objs/ibo.hpp"
+#include "../lib/rendering/gl_buffer_objs/vertexbufferlayout.hpp"
+
 #include "../lib/rendering/gl_buffer_objs/vao.hpp"
 
-#include "../lib/rendering/shapes/Triangle.h"
 #include "../lib/rendering/shapes/Rectangle.h"
+#include "../lib/rendering/shapes/Triangle.h"