+++ Project/gearworks/src/lib/rendering/gl_buffer_objs/vbo.hpp
@@ -24,11 +24,6 @@ public:
 		// Set the data of the VBO based on count and data
 		GL_CALL(glBufferData(GL_ARRAY_BUFFER, count * sizeof(float), data, GL_STATIC_DRAW));
 
-		// Enable the position attribute by enabling the attribute at index 0 (position is at index 0)
-		GL_CALL(glEnableVertexAttribArray(0));
-		// Specify how to read the position attribute (read it as positions in 2D space)
-		GL_CALL(glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE, 2 * sizeof(float), 0));
-
 		// Initially unbind the buffer so that it is bound through the VAO at render-time
 		Unbind();
 	}