+++ Project/gearworks/src/lib/utility/Utility.cpp
@@ -1,4 +1,4 @@
-#include "Utility.h"
+﻿#include "Utility.h"
 
 #include "gwrendering.h"
 
@@ -6,21 +6,29 @@
 
 // GLFW error callback function
 void util::glfwErrorCallback(int error, const char *description) {
+	// Send the error to console
 	std::cout << std::endl << "[GLFW] Error (" << error << "): " << description << std::endl;
 }
 
-// Current-bound VAO function
-int util::GetCurrentBoundVAO() {
-	int result;
-	glGetIntegerv(GL_VERTEX_ARRAY_BINDING, &result);
+unsigned int util::GetSizeOfType(unsigned int type) {
+	// Switch the type
+	switch (type) {
+		// The size of a float is 4 - if the type is GL_FLOAT, return 4
+		case GL_FLOAT: return 4;
+		// The size of an unsigned int is 4 - if the type is GL_UNSIGNED_INT, return 4
+		case GL_UNSIGNED_INT: return 4;
+		// The size of an unsigned char, or unsigned byte, is 1 - if the type is GL_UNSIGNED_BYTE, return 1
+		case GL_UNSIGNED_BYTE: return 1;
+	}
 
-	return result;
+	// If the type is invalid, return 0
+	return 0;
 }
 
 // coord::InitWorldCoordSystem
 void util::coord::InitWorldCoordSystem(unsigned int *program, float maxVal) {
 	// Initialize the coordinate system to world space.
-	// NOTE: The value '0.675' came about in trial and error. This is terrible practice and should be fixed at some point, but it works for now.
+	// NOTE: The value '0.675' came about in trial and error. This is terrible practice and should be fixed at some point, but it works for now. ¯\_(ツ)_/¯
 	// A conversion between world to screen coordinates or vice versa should also be made sooner or later.
 	glm::mat4 projMat = glm::ortho(-maxVal, maxVal, (float)-(maxVal * 0.675f), (float)(maxVal * 0.675f), -1.0f, 1.0f);
 	Shader::ModifyUniformmat4(program, "u_ModelViewProj", projMat);